bang: Bayesian Analysis, No Gibbs

Performs Bayesian analyses using some simple commonly-used models. The multivariate generalized ratio-of-uniforms method is used to simulate random samples from the required posterior distribution. The user can either choose hyperparameter values of a default prior distribution or specify their own prior distribution.

Details

Currently three conjugate hierarchical models are available: beta-binomial, gamma-Poisson and 1-way Analysis of Variance (ANOVA). The function hef produces random posterior samples from for the beta-binomial and gamma-Poisson models. The function hanova1 does this for the 1-way Analysis of Variance (ANOVA). The rust package is used to produce these samples.

See vignette("bang-a-vignette", package = "bang") for a brief introduction to the package and vignette("bang-b-hef-vignette", package = "bang") and vignette("bang-c-anova-vignette", package = "bang") for illustrations of the use of the hef and hanova1 functions.

References

Northrop, P. J. (2017). rust: Ratio-of-Uniforms Simulation with Transformation. R package version 1.2.3. https://cran.r-project.org/package=rust.

See also

Author

Maintainer: Paul J. Northrop p.northrop@ucl.ac.uk [copyright holder]

Authors:

Benjamin D. Hall benjamin.hall.16@ucl.ac.uk [copyright holder]