Weight Gained by Rats

Data from an experiment to study weight gained by 10 rats fed on four different diets, defined by a combination of the amount of protein (low and high) and by the source of protein (beef and cereal).

Usage

weight_gain

Format

A data frame with 40 rows and 3 columns.

Column 1, source: source of protein, a factor with levels Beef and Cereal.

Column 2, type: amount of protein, a factor with levels High and Low.

Column 3, weightgain: weight gained, in grams.

Source

D. J. Hand, A. D. Lunn, K. J. McConway, and E. Ostrowski (1994). A Handbook of Small Datasets, Chapman and Hall/CRC, London.