Summarizing hef objects

summary method for class "hef".

# S3 method for hef
summary(object, ..., params = c("hyper", "pop"),
  which_pop = 1:ncol(object$theta_sim_vals))

Arguments

object

an object of class "hef", a result of a call to hef.

...

Additional arguments passed on to summary.ru.

params

A character scalar.

If params = "hyper" then the posterior samples of all hyperparameter values in \(\phi\) are summarized using summary.ru.

If params = "pop" then only posterior samples of the populations specified in which_pop are summarized.

which_pop

An integer vector. If params = "pop" then which_pop indicates which populations, i.e. which columns of object$theta_sim_vals to summarize, using summary. The default is all populations.

Examples

# Beta-binomial model, rat data
rat_res <- hef(model = "beta_binom", data = rat)

# Posterior summaries of the hyperparameters alpha and beta
summary(rat_res)
#> ru bounding box:  
#>                box       vals1       vals2 conv
#> a        1.0000000  0.00000000  0.00000000    0
#> b1minus -0.2382163 -0.40313465 -0.03906170    0
#> b2minus -0.2174510  0.05447431 -0.35297539    0
#> b1plus   0.2231876  0.36718411 -0.06551353    0
#> b2plus   0.2512577  0.05665707  0.44459818    0
#> 
#> estimated probability of acceptance:  
#> [1] 0.5159959
#> 
#> sample summary 
#>      alpha              beta       
#>  Min.   : 0.9246   Min.   : 4.701  
#>  1st Qu.: 1.7489   1st Qu.:10.552  
#>  Median : 2.1809   Median :13.093  
#>  Mean   : 2.4250   Mean   :14.490  
#>  3rd Qu.: 2.8909   3rd Qu.:17.203  
#>  Max.   :10.4821   Max.   :58.499  

# Posterior summaries of the binomial probability for rats 1 to 3
summary(rat_res, params = "pop", which_pop = 1:3)
#>       p[1]               p[2]                p[3]          
#>  Min.   :0.001406   Min.   :0.0007488   Min.   :2.492e-05  
#>  1st Qu.:0.031375   1st Qu.:0.0297159   1st Qu.:3.247e-02  
#>  Median :0.053667   Median :0.0530726   Median :5.676e-02  
#>  Mean   :0.060860   Mean   :0.0622749   Mean   :6.331e-02  
#>  3rd Qu.:0.082465   3rd Qu.:0.0873819   3rd Qu.:8.690e-02  
#>  Max.   :0.210737   Max.   :0.2789649   Max.   :2.474e-01